authserver/src/requirements.txt
flask
PyJWT
cryptography
gunicorn
